Refactored ServerpodClientShared.callServerEndpoint() and subclasses to unify identical code
